@@ -9,7 +9,6 @@ import (
 	"github.com/gruntwork-io/terratest/modules/terraform"
 	testStructure "github.com/gruntwork-io/terratest/modules/test-structure"
 	"github.com/stretchr/testify/assert"
-	"k8s.io/apimachinery/pkg/util/runtime"
 )
 
 func TestExamplesMultipleSubnetsPerAZ(t *testing.T) {
@@ -37,10 +36,13 @@ func TestExamplesMultipleSubnetsPerAZ(t *testing.T) {
 	// At the end of the test, run `terraform destroy` to clean up any resources that were created
 	defer cleanup(t, terraformOptions, tempTestFolder)
 
-	// If Go runtime crushes, run `terraform destroy` to clean up any resources that were created
-	defer runtime.HandleCrash(func(i interface{}) {
-		cleanup(t, terraformOptions, tempTestFolder)
-	})
+	// If Go runtime panics, run `terraform destroy` to clean up any resources that were created
+	defer func() {
+		if r := recover(); r != nil {
+			cleanup(t, terraformOptions, tempTestFolder)
+			panic(r) // Re-panic after cleanup
+		}
+	}()
 
 	// This will run `terraform init` and `terraform apply` and fail the test if there are any errors
 	terraform.InitAndApply(t, terraformOptions)
@@ -61,17 +63,17 @@ func TestExamplesMultipleSubnetsPerAZ(t *testing.T) {
 	namedPrivateSubnetsStatsMap := terraform.OutputMapOfObjects(t, terraformOptions, "named_private_subnets_stats_map")
 	// Verify we're getting back the outputs we expect
 	assert.Equal(t, len(namedPrivateSubnetsStatsMap), 3)
-	assert.Equal(t, len(namedPrivateSubnetsStatsMap["backend"].([]map[string]any)), 2)
-	assert.Equal(t, len(namedPrivateSubnetsStatsMap["services"].([]map[string]any)), 2)
-	assert.Equal(t, len(namedPrivateSubnetsStatsMap["db"].([]map[string]any)), 2)
+	assert.Equal(t, len(namedPrivateSubnetsStatsMap["backend"].([]interface{})), 2)
+	assert.Equal(t, len(namedPrivateSubnetsStatsMap["services"].([]interface{})), 2)
+	assert.Equal(t, len(namedPrivateSubnetsStatsMap["db"].([]interface{})), 2)
 
 	// Run `terraform output` to get the value of an output variable
 	namedPublicSubnetsStatsMap := terraform.OutputMapOfObjects(t, terraformOptions, "named_public_subnets_stats_map")
 	// Verify we're getting back the outputs we expect
 	assert.Equal(t, len(namedPublicSubnetsStatsMap), 3)
-	assert.Equal(t, len(namedPublicSubnetsStatsMap["backend"].([]map[string]any)), 2)
-	assert.Equal(t, len(namedPublicSubnetsStatsMap["services"].([]map[string]any)), 2)
-	assert.Equal(t, len(namedPublicSubnetsStatsMap["db"].([]map[string]any)), 2)
+	assert.Equal(t, len(namedPublicSubnetsStatsMap["backend"].([]interface{})), 2)
+	assert.Equal(t, len(namedPublicSubnetsStatsMap["services"].([]interface{})), 2)
+	assert.Equal(t, len(namedPublicSubnetsStatsMap["db"].([]interface{})), 2)
 }
 
 func TestExamplesMultipleSubnetsPerAZDisabled(t *testing.T) {
